refactor(GameResult): mapToInt -> mapToLong 변경
- 4000000000원 같이 큰 수인 경우 연산 값이 안맞아서 수정

+++ src/main/java/lotto/model/GameResult.java
@@ -27,6 +27,6 @@ public class GameResult {
     }
 
     public double benefitResult(double deposit) {
-        return gameResult.stream().mapToInt(Rank::getWinningMoney).sum() / deposit;
+        return gameResult.stream().mapToLong(Rank::getWinningMoney).sum() / deposit;
     }
 }
